1.40.0-beta-20061108
git-svn-id: svn://tug.org/pdftex/trunk@70 7b5bf505-3498-49f7-8177-37ec8b90719f

+++ source/src/texk/web2c/pdftexdir/writettf.c
@@ -1083,7 +1083,7 @@ static void ttf_reindex_glyphs (void)
       append_new_glyph:
         assert (glyph > glyph_tab && glyph - glyph_tab < glyphs_count);
         if (glyph->newindex < 0) {
-            glyph_index[new_glyphs_count] = (short) (glyph - glyph_tab);
+            glyph_index[new_glyphs_count] =  glyph - glyph_tab;
             glyph->newindex = new_glyphs_count;
             new_glyphs_count++;
         }
